html/infrastructure/urls/terminology-0/document-base-url.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: document base URL</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/infrastructure.html#document-base-url">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/infrastructure.html#terminology-0">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  iframe { display: none }
</style>
<body onload="on_load()">
  <div id="log"></div>
  <script>

  var t1 = async_test("The document base URL must be the value of href attribute of the base element"),
      t2 = async_test("The document base URL must be document base URL of the container's"),
      t3 = async_test("The document base URL must be the base URL of the creator's"),
      t4 = async_test("The document base URL of an iframe srcdoc must be the document base URL of the container's");

  function on_load() {
    t1.step(function () {
      var base = document.createElement("base");
      base.setAttribute("href", "http://www.example.com/");
      document.head.appendChild(base);
      assert_equals(base.href, "http://www.example.com/", "The base.href should be 'http://www.example.com/'.");
      assert_equals(document.baseURI, base.href, "The document.baseURI should equal to base.href.");
    });
    t1.done();

    t2.step(function () {
      var fr = document.createElement("iframe");
      fr.setAttribute("src", "test.html");
      document.body.appendChild(fr);
      assert_equals(fr.contentDocument.baseURI, document.baseURI, "The baseURI should equal to the container's document base URL.");
    });
    t2.done();

    t3.step(function () {
      var fr = document.createElement("iframe");
      fr.setAttribute("src", "about:blank");
      document.body.appendChild(fr);
      assert_equals(fr.contentDocument.baseURI, document.baseURI, "The baseURI should equal to the creator's base URL.");
    });
    t3.done();

    t4.step(function () {
      var fr = document.createElement("iframe");
      fr.setAttribute("srcdoc", "<p>foobar</p>");
      document.body.appendChild(fr);
      assert_equals(fr.contentDocument.baseURI, document.baseURI, "The baseURI should be the container's document base URL.");
    });
    t4.done();
  }
  </script>
</body>
